Eliminar productos. Se pueden eliminar productos y se añadio ajax y swal alert, para preguntarle al usuario si esta seguro de eliminarlo

+++ src/main/java/sistemaInventario/SistemaInventario/repositorio/ProductoRepositorio.java
@@ -1,7 +1,3 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package sistemaInventario.SistemaInventario.repositorio;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -11,7 +7,6 @@ import sistemaInventario.SistemaInventario.modelo.Producto;
  *
  * @author joel_
  */
- 
-public interface ProductoRepositorio extends JpaRepository<Producto, Integer> {
-    // Puedes agregar métodos personalizados si es necesario
+public interface ProductoRepositorio extends JpaRepository<Producto, Long> {
+
 }
